Extract `ClickBehaviorSidebar` components (#25088)
* Extract straightforward styled components

* Extract `SidebarItem` bits

* Remove not used props

* Sort imports

* Extract `ValuesYouCanReference` component

* Extract `CustomLinkText` component

* Extract `LinkOption` component

* Extract `Column` component

* Extract utils

* Extract `TypeSelector` component

* Extract `ActionOptions` component

* Extract `QuestionDashboardPicker` component

* Extract `CrossfilterOptions` component

* Extract `LinkOptions` component

* Extract `TableClickBehaviorView` component

* Extract `ClickBehaviorSidebarHeader` component

* Extract `ClickBehaviorSidebarMainView` component

* Fix `this` references in functional components
